reduced to a single ui state
