+++ app/src/main/kotlin/edu/stanford/bdh/engagehf/bluetooth/data/models/BluetoothUiState.kt
@@ -1,8 +1,5 @@
 package edu.stanford.bdh.engagehf.bluetooth.data.models
 
-import edu.stanford.bdh.engagehf.messages.Message
-import edu.stanford.spezi.core.bluetooth.data.model.Measurement
-
 sealed interface BluetoothUiState {
     // Initial states
     data object Idle : BluetoothUiState
@@ -13,18 +10,4 @@ sealed interface BluetoothUiState {
 
     // Generic error
     data class Error(val message: String) : BluetoothUiState
-}
-
-data class MeasurementDialogUiState(
-    val measurement: Measurement? = null,
-    val isVisible: Boolean = false,
-    val isProcessing: Boolean = false,
-    val formattedWeight: String = "",
-)
-
-sealed interface Action {
-    data class ConfirmMeasurement(val measurement: Measurement) : Action
-    data object DismissDialog : Action
-    data class MessageItemClicked(val message: Message) : Action
-    data class ToggleExpand(val message: Message) : Action
 }
