+++ app/src/main/kotlin/edu/stanford/bdh/engagehf/bluetooth/data/models/UiState.kt
@@ -5,4 +5,6 @@ import edu.stanford.bdh.engagehf.messages.Message
 data class UiState(
     val vitalDisplay: VitalDisplayUiState = VitalDisplayUiState(),
     val messages: List<Message> = emptyList(),
+    val bluetooth: BluetoothUiState = BluetoothUiState.Idle,
+    val measurementDialog: MeasurementDialogUiState = MeasurementDialogUiState(),
 )
